Ensure weekends and holidays are correctly highlighted on the monthly calendar view
Fix issue where day of the week was not being properly cast to an integer, causing incorrect highlighting on the calendar view for weekends and holidays.

+++ src/Controllers/DiretorController.php
@@ -107,7 +107,7 @@ class DiretorController {
         $diasInfo = [];
         for ($d = 1; $d <= $diasNoMes; $d++) {
             $data = sprintf('%04d-%02d-%02d', $ano, $mes, $d);
-            $diaSemana = date('w', strtotime($data));
+            $diaSemana = (int)date('w', strtotime($data));
             $diasInfo[$d] = [
                 'data' => $data,
                 'diaSemana' => $diaSemana,
